refactor(package): 패키징 구조 변경
