+++ app/src/main/java/woowacourse/app/model/ticket/TicketMapper.kt
@@ -1,8 +1,7 @@
 package woowacourse.app.model.ticket
 
 import woowacourse.app.model.movie.MovieMapper.toUiModel
-import woowacourse.domain.ticket.Position
-import woowacourse.domain.ticket.Seat
+import woowacourse.app.model.seat.SeatMapper.toUiModel
 import woowacourse.domain.ticket.Ticket
 
 object TicketMapper {
@@ -11,34 +10,6 @@ object TicketMapper {
             movie = this.movie.toUiModel(),
             bookedDateTime = this.bookedDateTime,
             seat = this.seat.toUiModel(),
-        )
-    }
-
-    fun Seat.toUiModel(): SeatUiModel {
-        return SeatUiModel(
-            rank = this.rank,
-            position = this.position.toUiModel(),
-        )
-    }
-
-    fun SeatUiModel.toDomainModel(): Seat {
-        return Seat(
-            rank = this.rank,
-            position = this.position.toDomainModel(),
-        )
-    }
-
-    fun Position.toUiModel(): PositionUiModel {
-        return PositionUiModel(
-            row = this.row,
-            column = this.column,
-        )
-    }
-
-    fun PositionUiModel.toDomainModel(): Position {
-        return Position(
-            row = this.row,
-            column = this.column,
         )
     }
 }
